pyproject.toml
[project]
name = "chatbots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.128.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
    "uvicorn[standard]>=0.40.0",
]
